chore: fix linting
