+++ frontend/components/steps/CreateComputeBlockInformationStep.tsx
@@ -1,7 +1,7 @@
 import { AlertType, useAlert } from "@/hooks/useAlert"
 import { useGetComputeBlockInfoMutation } from "@/mutations/computeBlockMutation"
 import { IOType } from "@/components/CreateComputeBlockModal"
-import { type Entrypoint, type InputOutput, type ComputeBlockDraft } from "@/components/CreateComputeBlockModal"
+import { type PageProps, type Entrypoint, type InputOutput, type ComputeBlockDraft } from "@/components/CreateComputeBlockModal"
 import { useState } from "react"
 import LoadingAndError from "@/components/LoadingAndError"
 import Input from "@/components/inputs/Input"
@@ -15,21 +15,10 @@ const mapInputOutput = (data: InputOutput, type: IOType) => ({
   config: data.config || {},
 })
 
-export type PageProps = {
-  onNext: () => void,
-  onPrev?: () => void,
-  computeBlock?: ComputeBlockDraft,
-  setComputeBlock?: React.Dispatch<React.SetStateAction<ComputeBlockDraft>>,
-  setSelectedEntrypoint?: React.Dispatch<React.SetStateAction<Entrypoint | undefined>>,
-  selectedEntrypoint?: Entrypoint,
-  loading?: boolean,
-  projectName: string
-}
-
 export default function CreateComputeBlockInformationStep({
   onNext,
   setComputeBlock,
-  projectName
+  project
 }: PageProps) {
   const [computeBlockCustomName, setComputeBlockCustomName] = useState<string>("")
   const [repoURL, setRepoURL] = useState<string>("")
@@ -43,7 +32,7 @@ export default function CreateComputeBlockInformationStep({
     if (repoURL.length > 0) {
       const cb = await mutateAsync({
         compute_block_custom_name: computeBlockCustomName,
-        project_name: projectName,
+        project_name: project.name,
         cbc_url: repoURL,
       })
