--- frontend/hooks/useSelectedProject.tsx
@@ -1,34 +1,0 @@
-"use client"
-
-import type { Project } from "@/utils/types"
-import React, { createContext, useContext, useState, type PropsWithChildren } from "react"
-
-type SelectedProjectContextType = {
-  selectedProject: Project | undefined,
-  setSelectedProject: (project: Project | undefined) => void,
-}
-
-const initialState: SelectedProjectContextType = {
-  selectedProject: undefined,
-  setSelectedProject: () => { }
-}
-
-const SelectedProjectContext = createContext<SelectedProjectContextType>({
-  ...initialState
-})
-
-export function SelectedProjectProvider({ children }: PropsWithChildren) {
-  const [selectedProject, setSelectedProject] = useState<Project | undefined>(undefined)
-
-  return (
-    <SelectedProjectContext.Provider value={{
-      selectedProject,
-      setSelectedProject
-    }}
-    >
-      {children}
-    </SelectedProjectContext.Provider>
-  )
-}
-
-export const useSelectedProject = () => useContext(SelectedProjectContext)
